Небольшие изменения. Код вывожу, стобы видели работу программы
